+++ R/000.DEPRECATION.R
@@ -1,5 +1,16 @@
+isUseNamesNADefunct <- function() {
+  action <- if (getRversion() >= "4.4.0") "defunct" else "deprecated"
+  action <- getOption("matrixStats.useNames.NA", action)
+  action <- match.arg(action, choices = c("deprecated", "defunct"))
+  (action == "defunct")
+}
+
 deprecatedUseNamesNA <- function() {
-  .Deprecated(msg = "useNames = NA is deprecated. Instead, specify either useNames = TRUE or useNames = TRUE.", package = .packageName)
+  if (isUseNamesNADefunct()) {
+    .Defunct(msg = "useNames = NA is defunct. Instead, specify either useNames = TRUE or useNames = FALSE.", package = .packageName)
+  } else {
+    .Deprecated(msg = "useNames = NA is deprecated. Instead, specify either useNames = TRUE or useNames = FALSE.", package = .packageName)
+  }
 }