+++ R/zzz.R
@@ -13,6 +13,13 @@
     action <- match.arg(action, choices = c("deprecated", "defunct"))
     options(matrixStats.validateIndices = action)
   }
+
+  ## Deprecate/defunct useNames = NA /HB 2023-10-31
+  action <- Sys.getenv("R_MATRIXSTATS_USENAMES_NA", NA_character_)
+  if (!is.na(action)) {
+    action <- match.arg(action, choices = c("deprecated", "defunct"))
+    options(matrixStats.useNames.NA = action)
+  }
 }
 
 #' @useDynLib "matrixStats", .registration = TRUE, .fixes = "C_"