+++ tests/diff2_subset.R
@@ -16,7 +16,7 @@ for (setNames in c(TRUE, FALSE)) {
   for (l in 1:2) {
     for (d in 1:2) {
       for (idxs in index_cases) {
-        for (useNames in c(NA, TRUE, FALSE)) {
+        for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
           validateIndicesTestVector(x, idxs,
                                     ftest = diff2, fsure = diff2_R,
                                     lag = l, differences = d, useNames = useNames)