+++ tests/rowAllAnys.R
@@ -10,8 +10,7 @@ rowAlls_R <- function(x, value = TRUE, na.rm = FALSE, ..., useNames = NA) {
     dim <- dim(x) # for 0xN and Mx0 cases; needed in R (< 3.4.0)
     if (!isTRUE(all.equal(dim(y), dim))) {
       dim(y) <- dim
-      dimnames <- dimnames(x)
-      if (!is.null(dimnames)) dimnames(y) <- dimnames
+      dimnames(y) <- dimnames(x)
     }
     
     res <- apply(y, MARGIN = 1L, FUN = all, na.rm = na.rm)
@@ -30,8 +29,7 @@ rowAnys_R <- function(x, value = TRUE, na.rm = FALSE, ..., useNames = NA) {
     dim <- dim(x) # for 0xN and Mx0 cases; needed in R (< 3.4.0)
     if (!isTRUE(all.equal(dim(y), dim))) {
       dim(y) <- dim
-      dimnames <- dimnames(x)
-      if (!is.null(dimnames)) dimnames(y) <- dimnames
+      dimnames(y) <- dimnames(x)
     }
     
     res <- apply(y, MARGIN = 1L, FUN = any, na.rm = na.rm)
@@ -76,7 +74,7 @@ for (kk in 1:3) {
 
     for (na.rm in c(FALSE, TRUE)) {
       # Check names attribute
-      for (useNames in c(NA, TRUE, FALSE)) {
+      for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
         m0 <- rowAlls_R(x, na.rm = na.rm, useNames = useNames)
         m1 <- rowAlls(x, na.rm = na.rm, useNames = useNames)
         m2 <- colAlls(t(x), na.rm = na.rm, useNames = useNames)
@@ -122,7 +120,7 @@ for (setDimnames in c(TRUE, FALSE)) {
   dimnames(x) <- if (setDimnames) dimnames else NULL
   for (na.rm in c(FALSE, TRUE)) {
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
       r0 <- rowAlls_R(x, value = value, na.rm = na.rm, useNames = useNames)
       r1 <- rowAlls(x, value = value, na.rm = na.rm, useNames = useNames)
       r2 <- colAlls(t(x), value = value, na.rm = na.rm, useNames = useNames)
@@ -258,7 +256,7 @@ for (setDimnames in c(TRUE, FALSE)) {
   for (value in c("g", NA_character_)) {
     for (na.rm in c(FALSE, TRUE)) {
       # Check names attribute
-      for (useNames in c(NA, TRUE, FALSE)) {
+      for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
         r0 <- rowAlls_R(x, value = value, na.rm = na.rm, useNames = useNames)
         r1 <- rowAlls(x, value = value, na.rm = na.rm, useNames = useNames)
         r2 <- colAlls(t(x), value = value, na.rm = na.rm, useNames = useNames)
@@ -305,7 +303,7 @@ dimnames <- list(letters[1:3], LETTERS[1:3])
 for (setDimnames in c(TRUE, FALSE)) {
   dimnames(x) <- if (setDimnames) dimnames else NULL
   # Check names attribute
-  for (useNames in c(NA, TRUE, FALSE)) {
+  for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
     r0 <- rowAnys_R(x, value = 0, useNames = useNames)
     r1 <- rowAnys(x, value = 0, useNames = useNames)
     stopifnot(identical(r0, r1))