+++ tests/rowCumsums.R
@@ -7,10 +7,7 @@ rowCumsums_R <- function(x, ..., useNames = NA) {
   
   # Preserve dimnames attribute?
   dim(y) <- dim(x)
-  if (isTRUE(useNames)) {
-    dimnames <- dimnames(x)
-    if (!is.null(dimnames)) dimnames(y) <- dimnames      
-  }
+  dimnames(y) <- if (isTRUE(useNames)) dimnames(x) else NULL
   
   y
 }
@@ -37,7 +34,7 @@ for (mode in c("logical", "integer", "double")) {
       if (setDimnames) dimnames(x) <- dimnames
       else dimnames(x) <- NULL    
       # Check names attribute
-      for (useNames in c(NA, TRUE, FALSE)) {
+      for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
         # Row/column ranges
         r0 <- rowCumsums_R(x, useNames = useNames)
         r1 <- rowCumsums(x, useNames = useNames)
@@ -65,7 +62,7 @@ for (mode in c("logical", "integer", "double")) {
     if (setDimnames) dimnames(x) <- dimnames
     else dimnames(x) <- NULL    
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
       # Row/column ranges
       r0 <- rowCumsums_R(x, useNames = useNames)
       r1 <- rowCumsums(x, useNames = useNames)
@@ -139,7 +136,7 @@ for (mode in c("logical", "integer", "double")) {
     if (setDimnames) colnames(x) <- colnames
     else dimnames(x) <- NULL
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
       r0 <- rowCumsums_R(x, useNames = useNames)
       r1 <- rowCumsums(x, useNames = useNames)
       r2 <- t(rowCumsums(t(x), useNames = useNames))
@@ -158,7 +155,7 @@ for (mode in c("logical", "integer", "double")) {
     if (setDimnames) rownames(x) <- rownames
     else dimnames(x) <- NULL
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
       r0 <- rowCumsums_R(x, useNames = useNames)
       r1 <- rowCumsums(x, useNames = useNames)
       r2 <- t(rowCumsums(t(x), useNames = useNames))