+++ tests/rowCumsums_subset.R
@@ -7,10 +7,7 @@ rowCumsums_R <- function(x, ..., useNames = NA) {
   
   # Preserve dimnames attribute?
   dim(y) <- dim(x)
-  if (isTRUE(useNames)) {
-    dimnames <- dimnames(x)
-    if (!is.null(dimnames)) dimnames(y) <- dimnames      
-  }
+  dimnames(y) <- if (isTRUE(useNames)) dimnames(x) else NULL
 
   y
 }
@@ -33,7 +30,7 @@ for (setDimnames in c(TRUE, FALSE)) {
   for (rows in index_cases) {
     for (cols in index_cases) {
       # Check names attribute
-      for (useNames in c(NA, TRUE, FALSE)) {
+      for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
         validateIndicesTestMatrix(x, rows, cols,
                                   ftest = rowCumsums, fsure = rowCumsums_R, useNames = useNames)
         validateIndicesTestMatrix(x, rows, cols,