+++ tests/rowDiffs_subset.R
@@ -51,7 +51,8 @@ for (setDimnames in c(TRUE, FALSE)) {
       for (lag in 1:2) {
         for (differences in 1:3) {
           # Check dimnames attribute
-          useNames <- c(NA, TRUE, FALSE)[count %% 3 + 1]
+          useNames <- c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)
+          useNames <- useNames[count %% length(useNames) + 1]
           validateIndicesTestMatrix(x, rows, cols,
                                     ftest = rowDiffs, fsure = rowDiffs_R,
                                     lag = lag, differences = differences, useNames = useNames)