+++ tests/rowIQRs_subset.R
@@ -12,8 +12,7 @@ rowIQRs_R <- function(x, na.rm = FALSE, ..., useNames = NA) {
   
   # Preserve names attribute
   dim(q) <- c(2L, nrow(x))
-  names <- rownames(x)
-  if (isTRUE(useNames) && !is.null(names)) colnames(q) <- names
+  colnames(q) <- if (isTRUE(useNames)) rownames(x) else NULL
   
   q[2L, , drop = TRUE] - q[1L, , drop = TRUE]
 }
@@ -48,7 +47,8 @@ for (setDimnames in c(TRUE, FALSE)) {
     for (cols in index_cases) {
       count <- count + 1L
       na.rm <- c(TRUE, FALSE)[count %% 2 + 1]
-      useNames <- c(NA, TRUE, FALSE)[count %% 3 + 1]
+      useNames <- c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)
+      useNames <- useNames[count %% length(useNames) + 1]
       
       validateIndicesTestMatrix(x, rows, cols,
                                 ftest = rowIQRs, fsure = rowIQRs_R,