+++ tests/rowOrderStats.R
@@ -82,7 +82,7 @@ for (setDimnames in c(TRUE, FALSE)) {
   if (setDimnames) dimnames(x) <- dimnames
   else dimnames(x) <- NULL
   # Check names attribute
-  for (useNames in c(NA, TRUE, FALSE)) {
+  for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
     y0 <- rowOrderStats_R(x, probs = probs, useNames = useNames)
     y1 <- rowOrderStats(x, which = which, useNames = useNames)
     stopifnot(all.equal(y1, y0))