+++ tests/rowWeightedMeans_subset.R
@@ -58,7 +58,8 @@ for (mode in c("numeric", "integer", "logical")) {
       for (cols in index_cases) {
         count <- count + 1L
         na.rm <- c(TRUE, FALSE)[count %% 2 + 1]
-        useNames <- c(NA, TRUE, FALSE)[count %% 3 + 1]
+        useNames <- c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)
+        useNames <- useNames[count %% length(useNames) + 1]
 
         validateIndicesTestMatrix_w(x, w, rows, cols, 
                                     ftest = rowWeightedMeans, fsure = rowWeightedMeans_R,