+++ NEWS.md
@@ -1,3 +1,16 @@
+# Version 1.1.0 [2023-11-06]
+
+## Deprecated and Defunct
+
+ * `useNames = NA` is defunct in R (>= 4.4.0). Remains deprecated in
+   R (< 4.4.0) for now.
+
+## Miscellaneous
+
+ * The deprecation warning for using `useNames = NA`, suggested using
+   `useNames = TRUE` twice instead of also `useNames = FALSE`.
+ 
+
 # Version 1.0.0 [2023-06-01]
 
 ## Significant Changes